@@ -80,11 +80,11 @@
     "lodash.uniq": "^4.5.0",
     "minimist": "1.2.0",
     "normalize.css": "^8.0.0",
-    "prop-types": "^15.6.0",
+    "prop-types": "^15.6.1",
     "query-string": "^5.0.0",
     "ramda": "^0.25.0",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
+    "react": "16.4.0",
+    "react-dom": "16.4.0",
     "react-helmet": "5.2.0",
     "react-hot-loader": "4.0.0",
     "react-intl": "^2.4.0",
@@ -129,50 +129,50 @@
     "chai-jest-diff": "^1.0.1",
     "cheerio": "1.0.0-rc.2",
     "circular-dependency-plugin": "4.4.0",
-    "css-loader": "^0.28.7",
+    "css-loader": "^0.28.11",
     "enzyme": "^3.1.0",
     "enzyme-adapter-react-16": "^1.0.2",
-    "eslint": "4.18.0",
+    "eslint": "4.19.1",
     "eslint-config-apptension-react": "^0.0.10",
-    "eslint-import-resolver-webpack": "0.8.4",
+    "eslint-import-resolver-webpack": "0.10.0",
     "eslint-plugin-chai-friendly": "^0.4.0",
-    "eslint-watch": "^3.1.2",
+    "eslint-watch": "^3.1.5",
     "eventsource-polyfill": "0.9.6",
     "exports-loader": "0.7.0",
-    "favicons-webpack-plugin": "^0.0.7",
-    "file-loader": "1.1.6",
+    "favicons-webpack-plugin": "^0.0.9",
+    "file-loader": "1.1.11",
     "html-loader": "0.5.5",
     "html-webpack-plugin": "^2.30.1",
     "husky": "^0.14.3",
-    "image-webpack-loader": "4.1.0",
-    "imports-loader": "0.7.1",
+    "image-webpack-loader": "4.2.0",
+    "imports-loader": "0.8.0",
     "inquirer-directory": "^2.1.0",
     "isomorphic-fetch": "^2.2.1",
     "jest": "^22.3.0",
     "jest-enzyme": "^4.2.0",
     "jest-junit": "^3.1.0",
     "jest-styled-components": "^4.10.0",
-    "json-server": "^0.12.1",
+    "json-server": "^0.12.2",
     "lint-staged": "6.1.1",
     "ngrok": "2.2.26",
-    "nock": "^9.1.6",
+    "nock": "^9.2.6",
     "node-plop": "0.9.0",
     "null-loader": "0.1.1",
     "offline-plugin": "4.9.0",
     "plop": "^1.9.1",
-    "react-test-renderer": "^16.0.0",
+    "react-test-renderer": "^16.4.0",
     "redux-saga-tester": "^1.0.197",
     "rimraf": "2.6.2",
-    "shelljs": "0.8.1",
+    "shelljs": "0.8.2",
     "sinon": "^4.3.0",
     "sinon-chai": "^2.14.0",
-    "style-loader": "^0.20.2",
+    "style-loader": "^0.21.0",
     "url-loader": "0.6.2",
     "webpack": "3.11.0",
     "webpack-appversion-plugin": "^1.2.1",
     "webpack-dev-middleware": "2.0.5",
-    "webpack-hot-middleware": "2.21.0",
-    "webpack-spritesmith": "^0.4.0"
+    "webpack-hot-middleware": "2.22.2",
+    "webpack-spritesmith": "^0.4.1"
   },
   "jest": {
     "setupTestFrameworkScriptFile": "./internals/testing/test.index.js",
