@@ -21,13 +21,13 @@ export class Home extends PureComponent {
     }).isRequired,
   };
 
-  componentWillMount() {
+  componentDidMount() {
     this.props.fetchMaintainers(this.props.language);
   }
 
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.language !== this.props.language) {
-      this.props.fetchMaintainers(nextProps.language);
+  componentDidUpdate(prevProps) {
+    if (prevProps.language !== this.props.language) {
+      this.props.fetchMaintainers(this.props.language);
     }
   }
 
